move getLabelSelector
