+++ pkg/operator/client/manifest_delete_test.go
@@ -5,6 +5,8 @@ import (
 	"testing"
 
 	"github.com/replicatedhq/kots/pkg/operator/applier"
+	"github.com/stretchr/testify/require"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime/schema"
@@ -948,6 +950,49 @@ func Test_hasResources(t *testing.T) {
 			if got := hasResources(tt.args.resourcesMap); got != tt.want {
 				t.Errorf("hasResources() = %v, want %v", got, tt.want)
 			}
+		})
+	}
+}
+
+func Test_getLabelSelector(t *testing.T) {
+	tests := []struct {
+		name             string
+		appLabelSelector metav1.LabelSelector
+		want             string
+	}{
+		{
+			name: "no requirements",
+			appLabelSelector: metav1.LabelSelector{
+				MatchLabels:      nil,
+				MatchExpressions: nil,
+			},
+			want: "",
+		},
+		{
+			name: "one requirement",
+			appLabelSelector: metav1.LabelSelector{
+				MatchLabels: map[string]string{
+					"kots.io/label": "abc",
+				},
+				MatchExpressions: nil,
+			},
+			want: "kots.io/label=abc",
+		},
+		{
+			name: "two requirements",
+			appLabelSelector: metav1.LabelSelector{
+				MatchLabels: map[string]string{
+					"kots.io/label": "abc",
+					"otherlabel":    "xyz",
+				},
+				MatchExpressions: nil,
+			},
+			want: "kots.io/label=abc,otherlabel=xyz",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			require.Equal(t, tt.want, getLabelSelector(&tt.appLabelSelector))
 		})
 	}
 }
